Change Policy of showing comments
All comments are displayed inside the page let.
Moves Comment cell in a separate file.
