+++ IOS/show-me-aui/show-me-aui/PageletView.swift
@@ -9,51 +9,6 @@
 import UIKit
 import SnapKit
 
-class CommentCell: UITableViewCell {
-  let userImageView = UIImageView()
-  let userNameLabel = UILabel()
-  let commentLabel = UILabel()
-  
-  override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
-    super.init(style: style, reuseIdentifier: reuseIdentifier)
-    
-    userImageView.layer.cornerRadius = 15
-    userImageView.clipsToBounds = true
-    self.contentView.addSubview(userImageView)
-    
-    userNameLabel.textColor = UIColor(red: 45, green: 96, blue: 0)
-    userNameLabel.font = UIFont(name: "Chalkduster", size: 12)
-    self.contentView.addSubview(userNameLabel)
-    
-    commentLabel.numberOfLines = 0
-    commentLabel.font = UIFont(name: "GillSans", size: 14)
-    commentLabel.textColor = UIColor(red: 21, green: 45, blue: 0)
-    self.contentView.addSubview(commentLabel)
-    
-    userImageView.snp.makeConstraints { make in
-      make.top.left.equalTo(self.contentView).offset(8)
-      make.width.equalTo(30)
-      make.height.equalTo(userImageView.snp.width)
-    }
-    
-    userNameLabel.snp.makeConstraints { make in
-      make.top.equalTo(contentView).offset(8)
-      make.left.equalTo(userImageView.snp.right).offset(8)
-      make.right.equalTo(contentView).offset(-4)
-    }
-    
-    commentLabel.snp.makeConstraints { make in
-      make.top.equalTo(userNameLabel.snp.bottom).offset(4)
-      make.bottom.right.equalTo(contentView).offset(-4)
-      make.left.equalTo(userImageView.snp.right).offset(8)
-    }
-  }
-  
-  required init?(coder aDecoder: NSCoder) {
-    super.init(coder: aDecoder)
-  }
-}
-
 class PageletView: UIView {
   private let userImageView = UIImageView()
   private let addressLabel = UILabel()
@@ -66,6 +21,7 @@ class PageletView: UIView {
   private let commentsTableView = UITableView()
   private let horizontalLine = UIView()
   
+  // Id of the current image pagelet.
   let imageId: Int
   
   init?(imageId: Int) {
@@ -73,12 +29,15 @@ class PageletView: UIView {
     super.init(frame: CGRect.zero)
   }
   
-  // Height of the comments (calcualtes the height for at most 3) as only 3 are show at
-  // a time.
-  var commentTablePreviewHeight: CGFloat {
+  required init?(coder aDecoder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+  
+  // Height of the comments table.
+  var currentCommentTableHeight: CGFloat {
     get {
       var height = CGFloat(0)
-      for row in 0..<min(3,commentsTableView.numberOfRows(inSection: 0)) {
+      for row in 0..<commentsTableView.numberOfRows(inSection: 0) {
         let cellPath = IndexPath(row: row, section: 0)
         if let cell = commentsTableView.cellForRow(at: cellPath) {
           height += cell.frame.height
@@ -91,14 +50,10 @@ class PageletView: UIView {
   }
   
   // Height of the pageLetView.
-  var currentHeight: CGFloat {
+  var currentPageLetHeight: CGFloat {
     get {
-      return commentsTableView.frame.origin.y + commentTablePreviewHeight
+      return commentsTableView.frame.origin.y + currentCommentTableHeight
     }
-  }
-  
-  required init?(coder aDecoder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
   }
   
   init(userImage uimage: UIImage, userName name: String, pageletImage pimage: UIImage) {
@@ -162,17 +117,19 @@ class PageletView: UIView {
     
     setupThumbDown()
     setUpCommentButton()
-    setUpCommentsTableView()
     
     self.addSubview(horizontalLine)
     horizontalLine.snp.makeConstraints { make in
       make.top.equalTo(likeButton.snp.bottom).offset(8)
-      make.left.right.bottom.equalTo(self)
+      make.left.right.equalTo(self)
     }
+    
+    setUpCommentsTableView()
   }
   
   func setUpCommentsTableView() {
-    commentsTableView.isHidden = true
+    commentsTableView.allowsSelection = false
+    commentsTableView.separatorStyle = .none
     commentsTableView.estimatedRowHeight = 60
     commentsTableView.rowHeight = UITableViewAutomaticDimension
     
@@ -180,6 +137,11 @@ class PageletView: UIView {
     commentsTableView.dataSource = self
     
     self.addSubview(commentsTableView)
+    commentsTableView.snp.makeConstraints { make in
+      make.top.equalTo(horizontalLine.snp.bottom).offset(8)
+      make.left.right.bottom.equalTo(self)
+      make.height.equalTo(currentCommentTableHeight)
+    }
   }
   
   func setUpCommentButton() {
@@ -284,19 +246,6 @@ class PageletView: UIView {
   
   func didTapComment() {
     print("didTapComment")
-    
-    commentsTableView.isHidden = false
-    
-    horizontalLine.snp.remakeConstraints { make in
-      make.top.equalTo(likeButton.snp.bottom).offset(8)
-      make.left.right.equalTo(self)
-    }
-    
-    commentsTableView.snp.remakeConstraints { make in
-      make.top.equalTo(horizontalLine.snp.bottom).offset(8)
-      make.left.right.bottom.equalTo(self)
-      make.height.equalTo(commentTablePreviewHeight)
-    }
   }
 }
 
@@ -337,23 +286,5 @@ extension PageletView: UITableViewDataSource, UITableViewDelegate {
     }
     
     return cell
-  }
-}
-
-extension UIColor {
-  convenience init(red: Int, green: Int, blue: Int) {
-    assert(red >= 0 && red <= 255, "Invalid red component")
-    assert(green >= 0 && green <= 255, "Invalid green component")
-    assert(blue >= 0 && blue <= 255, "Invalid blue component")
-    
-    self.init(red: CGFloat(red) / 255.0, green: CGFloat(green) / 255.0, blue: CGFloat(blue) / 255.0, alpha: 1.0)
-  }
-  
-  convenience init(rgb: Int) {
-    self.init(
-      red: (rgb >> 16) & 0xFF,
-      green: (rgb >> 8) & 0xFF,
-      blue: rgb & 0xFF
-    )
   }
 }
